monitoring-center-backend/test-requirements.txt
pytest
pylint
pytest-cov
pytest-pycodestyle
pytest-isort
pytest-mccabe
pytest-mypy
pytest-pylint
pytest-flask
pytest-flask-sqlalchemy
